Changed namespaces to match renamed sink projects
